requirements.txt
anyascii==0.1.7
appdirs==1.4.3
beautifulsoup4==4.8.2
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
Django==2.2.19
django-extensions==3.1.1
django-filter==2.4.0
django-modelcluster==5.1
django-taggit==1.3.0
django-tailwind==1.1.1
django-treebeard==4.5.1
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
draftjs-exporter==2.1.7
et-xmlfile==1.0.1
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
l18n==2020.6.1
lockfile==0.12.2
msgpack==0.6.2
openpyxl==3.0.7
packaging==20.3
pep517==0.8.2
# Pillow==7.2.0
progress==1.5
# psycopg2==2.8.6
pyparsing==2.4.6
pytoml==0.1.21
pytz==2021.1
requests==2.22.0
retrying==1.3.3
six==1.14.0
soupsieve==2.2.1
sqlparse==0.4.1
tablib==3.0.0
Unidecode==1.2.0
urllib3==1.25.8
wagtail==2.12.3
wagtailmedia==0.7.0
webencodings==0.5.1
Willow==1.4
xlrd==2.0.1
XlsxWriter==1.3.8
xlwt==1.3.0
